tests, net, kmp: Add client arg
As of its next release, openshift-python-wrapper will enforce passing
client when working with cluster resources.
openshift-virtualization-tests must align with this change.
All calls in the code to openshift-python-wrapper resources should be
updated to pass client arg.

Since the KMP NS is created with an admin_client the VMs must be created
with the same client - VMs under a namespace that is not created by a project
admin require admin privileges.

Signed-off-by: Anat Wax <awax@redhat.com>