[4.20] [upgrade/network] Remove redundant network upgrade tests
Removing 3 network upgrade tests that are not contributing anything.
The dependent fixtures and functions were also removed, except for those that are
still used by other flows.
Polarion test cases deleted as well.

The rationale of removing each of these tests:
- test_nmstate_bridge_before_upgrade: Needless to explain this is a pure nmstate flow, that
doesn't involve CNV.
- test_bridge_marker_before_upgrade: Bridge-marker is indirectly covered in any bridge test,
which would fail if the bridge-marker is disfunctioning.
- test_install_ovs_with_annotations_before_upgrade: Actively opting-in OVS CNI is not needed
on OVN-k8s clusters, which is now the default.