CherryPicked: [cnv-4.20] Remove Artifactory dependency from CPU affinity test (#3247)
The test for kubevirt_vmi_node_cpu_affinity is flaky duo to connectivity
issues with the artifactory, in this PR I modified the test to not rely
on the artifactory to avoid this kind of failures to stabilize the
observability lanes.

https://issues.redhat.com/browse/CNV-75209

##### Short description:

##### More details:

##### What this PR does / why we need it:

##### Which issue(s) this PR fixes:

##### Special notes for reviewer:

##### jira-ticket:
<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->