net, udn: Passt migration connectivity test
Introduce passt based seamless migration tests
similar to those that are currently testing
l2bridge binding.

Move common udn fixtures to conftest.py

wait_for_ready_vm_with_restart:

In order to be able to wait for external dependencies
to materialize without specifically checking them,
add a function that waits for READY condition.
If the condition is not met within the time period
restart the VM and wait again for the condition.
Restarting the VM is required in scenarios that
VM/VMI does not know how to wait for dependencies
and is stuck in either pending or failed state.

Assited by Cursor AI.

Signed-off-by: Nir Dothan <ndothan@redhat.com>